rm size_t

+++ include/http/request.h
@@ -97,7 +97,7 @@ void http_request_set_payload(http_request_t *req, char *payload);
 void http_response_free(http_response_t *res);
 void http_response_add_header(http_response_t *req, const char *key, const char *value);
 void http_response_set_content_type(http_response_t *req, const char *value);
-void http_response_set_payload(http_response_t *req, char *payload, size_t size);
-void http_response_add_payload(http_response_t *req, char *payload, size_t size);
+void http_response_set_payload(http_response_t *req, char *payload, uint32_t size);
+void http_response_add_payload(http_response_t *req, char *payload, uint32_t size);
 
 #endif /* HTTPCLIENT_REQUEST_H_ */
